图片选择上传 日期选择

+++ app/src/main/java/com/hdcy/app/model/Bean4VedioDetail.java
@@ -20,7 +20,7 @@ public class Bean4VedioDetail implements Serializable {
     public  int viewCount;
     public  boolean live;
     public  boolean start;
-    public  int startTime;
+    public  long startTime;
     public  String length;
     public  int commentCount;
